// generated/src/app/api-models/letter-print-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class LetterPrintSetup {
  seqLtprtId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  requestType: string;
  action: string;
  letterType: string;
  printBatch: string;
  sortOrder: string;
  originalPrintDate: Date;
  numCopies: number;
  status: string;
  daemonRequest: string;
  template: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  reprintRequest: string;
  inProcess: string;
}